requirements.txt
python-telegram-bot==20.7
telethon==1.34.0
httpx==0.25.2
beautifulsoup4==4.12.3
apscheduler==3.10.4
aiosqlite==0.20.0
deep-translator==1.11.4
openai==1.35.0
python-dotenv==1.0.1
pytz==2024.1
lxml==5.2.2
